+++ scripts/libnet_examples.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+. libnet.sh
+
+get_free_tcp_port