+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 - bpf: test gcc toolchain;
 #### Added
+- scripts: add libnet.sh;
 - glibc: test bcmp() and bzero(), memccpy();
 - version: config.json: add trace-cmd;
 - bpf: add verifier subdir;